[refactor] #88 transaction 메소드 수정
